+++ lib/src/main/kotlin/kotlinDactyl/client/files/models/ClientFileModel.kt
@@ -1,12 +1,13 @@
 package kotlinDactyl.client.files.models
 
+import kotlinDactyl.helpers.DataTransformers.SizeTransform
 import java.time.OffsetDateTime
 
 data class ClientFileModel(
     val name:String,
     val mode:String,
     val modeBits:String,
-    val size:Int,
+    val size:Long,
     val isFile:Boolean,
     val isSymlink:Boolean,
     val mimeType:String,
